requirements.txt
pyinstaller
requests
websocket-client==1.5.1
websocket-server==0.6.4
Pillow
Flask
beautifulsoup4